Conflicts:
	GSStravel/src/main/java/controller/TravelServlet.java

+++ GSStravel/src/main/webapp/QandAResponse.jsp
@@ -18,6 +18,14 @@
 	src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"
 	integrity="sha384-Tc5IQib027qvyjSMfHjOMaLkfuWVxZxUPnCJA7l2mCWNIpG9mGCD8wGNIcPD7Txa"
 	crossorigin="anonymous"></script>
+<style type="text/css">
+.color-red{
+	border-color:red
+}
+.color-green{
+	border-color:green
+}
+</style>
 <title>回應</title>
 <style>
 textarea {
@@ -48,48 +56,48 @@ textarea {
 						<td colspan="2"><textarea readonly>${list.question_Text}</textarea></td>
 					</tr>
 				</table>
-				<form action='<c:url value="/QandAInsertServlet"/>'>
-					<input type="hidden" name="qa_No" value="${list.qa_No}"> <input
-						type="hidden" name="answer_No"
-						value="<%=session.getAttribute("emp_No")%>">
+				<form action='<c:url value="/QandAInsertServlet?role=true"/>' method="POST">
+					<input type="hidden" name="qa_No" value="${list.qa_No}">
+					<input type="hidden" name="answer_No" value="<%=session.getAttribute("emp_No")%>">
 					<c:if test="${list.answer_No!=0}">
 						<table>
 							<h2>回應</h2>
 							<tr>
-								<td><textarea name="answer_Text" class="Ans_textarea"
-										readonly="readonly">${list.answer_Text}</textarea></td>
+								<td><textarea name="answer_Text" class="Ans_textarea" readonly="readonly">${list.answer_Text}</textarea></td>
 							</tr>
 						</table>
 					</c:if>
 					<br />
-					<c:if test="${emp_Role eq true}">
+					<c:if test="${role}">
 						<c:if test="${list.answer_No==0}">
 							<div class="displayclass">
 								<table>
 									<tr>
 										<td>回應內容</td>
-										<td><textarea name="answer_Text"></textarea></td>
+										<td><textarea class="answer_Text" name="answer_Text"></textarea></td>
 									</tr>
 								</table>
 							</div>
-							<button type="button" class="notdisplayclass" onclick="Resopse()">我要回應</button>
-							<button type="submit" name="prodaction" value="insertAnswer"
-								class="displayclass">回應</button>
+							<button type="button" class="notdisplayclass" onclick="Respose()">我要回應</button>
+							<button type="submit" name="prodaction" value="insertAnswer" class="displayclass displaybb">回應</button>
+							<button type="button" class="displayclass" onclick="cancelrespose()">取消</button>
+							<button type="button" class="notdisplayclass" onclick="window.location = '/GSStravel/QandAServlet?role=true';">回上一頁</button>
 						</c:if>
 					</c:if>
-
-					<c:if test="${(emp_Role eq true)&&(list.answer_No!=0)}">
-						<button type="button" class="notdisplaybutton"
-							onclick="updateData()">修改</button>
-						<button type="submit" class="displaybutton" name="prodaction"
-							value="updateAnswer">確認修改</button>
+					<c:if test="${(emp_Role eq true)&&(list.answer_No!=0)&&role}">
+						<button type="button" class="notdisplaybutton" onclick="updateData()">修改</button>
+						<button type="submit" class="displaybutton displayaa" name="prodaction" value="updateAnswer">確認修改</button>
 						<button type="button" class="displaybutton" onclick="cancel()">取消修改</button>
-						<button type="button" class="notdisplaybutton"
-							onclick="checkdelete()">刪除</button>
+						<button type="button" class="notdisplaybutton" onclick="checkdelete()">刪除</button>
+						<button type="button" class="notdisplaybutton" onclick="window.location = '/GSStravel/QandAServlet?role=true';">回上一頁</button>
 					</c:if>
 				</form>
-				<button type="button"
-					onclick="window.location = '/GSStravel/QandAServlet';">回上一頁</button>
+				<c:if test="${!role}">
+					<c:if test="${list.question_No==emp_No}">
+						<button type="button" class="notdisplaybutton" onclick="checkdelete()">刪除</button>
+					</c:if>
+					<button type="button" onclick="window.location = '/GSStravel/QandAServlet?role=false';">回上一頁</button>
+				</c:if>
 			</div>
 		</div>
 	</div>
@@ -98,10 +106,16 @@ textarea {
 	$(".displayclass").hide();
 	$(".displaybutton").hide();
 	var Ans_textarea;
-
-	function Resopse() {
+	function Respose() {
 		$(".displayclass").show();
 		$(".notdisplayclass").hide();
+	}
+	function cancelrespose() {
+		$(".answer_Text").val('');
+		$(".displayclass").hide();
+		$(".notdisplayclass").show();
+		$(".answer_Text").removeClass("color-green");
+		$(".answer_Text").removeClass("color-red");
 	}
 	function updateData() {
 		Ans_textarea = $(".Ans_textarea").val();
@@ -114,17 +128,38 @@ textarea {
 		$(".displaybutton").hide();
 		$(".notdisplaybutton").show();
 		$(".Ans_textarea").prop("readonly", "readonly");
+		$(".Ans_textarea").removeClass("color-green");
+		$(".Ans_textarea").removeClass("color-red");
+		
 	}
-
 	function checkdelete() {
 		if (confirm("確定要刪除?")) {
-			window.location = '/GSStravel/QandAInsertServlet?prodaction=deleteOne&qa_No='
-					+ $
-			{
-				list.qa_No
-			}
-			;
+			window.location = '/GSStravel/QandAInsertServlet?prodaction=deleteOne&qa_No='+ ${list.qa_No}+'&role='+${role};
 		}
 	}
+	$(".answer_Text").blur(function(){
+		$(".answer_Text").removeClass("color-green");
+		$(".answer_Text").removeClass("color-red");
+		if($(".answer_Text").val()==''){
+			$(".answer_Text").addClass("color-red");
+			$(".displaybb").prop("disabled", true);
+		}
+		else{
+			$(".answer_Text").addClass("color-green");	
+			$(".displaybb").prop("disabled", false);
+		}		
+	});
+	$(".Ans_textarea").blur(function(){
+		$(".Ans_textarea").removeClass("color-green");
+		$(".Ans_textarea").removeClass("color-red");
+		if($(".Ans_textarea").val()==''){
+			$(".Ans_textarea").addClass("color-red");
+			$(".displayaa").prop("disabled", true);
+		}
+		else{
+			$(".Ans_textarea").addClass("color-green");	
+			$(".displayaa").prop("disabled", false);
+		}		
+	});
 </script>
 </html>
